+++ footer.php
@@ -20,12 +20,22 @@ if (file_exists($bh_dir."/myFooter.php")) {
 } elseif (dynamic_sidebar('footer-simple')) {
 // sidebar area
 } else {
+
+// CHECK: ClassicPress or WordPress
+global $cp_version;
+if (!$cp_version) {
+	$cmsLink = '<a href="https://wordpress.org">WordPress</a>';
+} else {
+	$cmsLink = '<a href="https://classicpress.net">ClassicPress</a>';
+}
+
 ?>
 
 <div class="container-fluid">
 <hr class="footer-thin">
 	<div class="text-center center-block">
-<p class="txt-railway"><a href="https://elite-star-services.com/blue-haze/">Blue Haze</a> for <a href="https://wordpress.org">WordPress</a> 
+	
+<p class="txt-railway"><a href="https://elite-star-services.com/blue-haze/">Blue Haze</a> for <?php echo $cmsLink; ?> 
 | © <?php echo date("Y"); ?> <a href="https://elite-star-services.com">Elite Star Services</a></p>
 	</div>
 </div>